Fix issue with the user code format (#43)
If the user code is specified in the YAML configuration as an
all-digit value without quotes, it is parsed as an integer,
except that we expect to work on a string. This fixes that by
making sure that as long as the value we read is not empty,
we make sure we will store it as a string.