✨ [feature] Add support for `bypass` parameter when arming
The `ARMING` action of the panel does support a `bypass`
parameter which allows to bypass the open sensors when
arming. This introduces the support for that parameter
in the configuration, as well as the ability to use the
`ARM_CUSTOM_BYPASS` service in Home Assistant to arm with
the `bypass` enabled.