🐛 [bugfix] Control commands weren't reading the `delay` and `bypass` parameters (#66)
We had support for overriding `delay` and `bypass` when receiving a
control command, which could be done in an Home Assistant automation
for instance to have more control over what `qolsysgw` supports but
that might not be exposable directly to Home Assistant. However,
the parameters received from the control command were never read,
and thus simply ignored. This fixes that.