✨ [feature] disable by default entities that do not update (#72)
Entities such as 'bluetooth' devices are declared as sensors but do
not seem to receive updates from the panel after that. We can thus
simply disable those by default, allowing people that want to see
them to enable them on a sensor-by-sensor basis, or even enable
all of them thanks to a new configuration option
`enable_static_sensors_by_default` that can be set to `true`.