✨ [feature] Add support for tampered sensors
This adds a `tampered` attribute to sensors indicating if the
sensor is currently believed to be tampered with or not.

This is not 100% accurate as it depends on specific sets of events
that the panel is sending in a given order, which allow to guess
if a sensor is currently tampered or not. This should be taken with
a huge grain of salt as, for instance, we do not have that
information when getting the summary state for the sensors from the
panel (and will be able to know a sensor was tampered when it won't
be tampered anymore).

More details are available at https://github.com/XaF/qolsysgw/issues/73
as to the identified interactions and how they are implemented here.