🐛 [bugfix] handle properly when to require a user code from HA (#81)
The gateway was expecting a user code from Home Assistant as soon
as a code was required for the action to be performed. However,
when a panel code is defined in the gateway configuration, this
code might not be required if not specifically required by the
`code_*_required` flags. This fixes that.