🐛 Sensor unique IDs deduplication when using modules
Some sensors like TakeoverModules and Translators will create extra
sensors which share exactly the same sensor ID as the module. This
makes sure that every sensor has a unique ID by adding back the zone
ID after the sensor ID when the sensor is not the first one which
that ID. This will ensure that we can still keep cross-installation
sensors for the main sensors, but handle deduplication when using
such modules.

Fixes https://github.com/XaF/qolsysgw/issues/108